# docs/api/backend_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TFG Chatbot Backend/Gateway API",
    "version": "0.1.0",
    "description": "API Gateway for the TFG Chatbot. Handles authentication, session management, and routes requests to internal services."
  },
  "paths": {
...
    "/professor/subjects": {
      "get": {
        "tags": [
          "professor"
        ],
        "summary": "List Professor Subjects",
        "description": "List all subjects the professor teaches with student counts.\n\nFor each subject, includes the number of enrolled students and documents.",
        "operationId": "list_professor_subjects_professor_subjects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SubjectInfo"
                  },
                  "type": "array",
                  "title": "Response List Professor Subjects Professor Subjects Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/professor/subjects/{subject}/students": {
      "get": {
        "tags": [
          "professor"
        ],
        "summary": "List Subject Students",
        "description": "List all students enrolled in a specific subject.\n\nOnly professors who teach this subject can view the student list.",
        "operationId": "list_subject_students_professor_subjects__subject__students_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudentInfo"
                  },
                  "title": "Response List Subject Students Professor Subjects  Subject  Students Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/professor/subjects/{subject}/documents": {
      "get": {
        "tags": [
          "professor"
        ],
        "summary": "List Subject Documents",
        "description": "List all documents for a specific subject from the RAG service.",
        "operationId": "list_subject_documents_professor_subjects__subject__documents_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentInfo"
                  },
                  "title": "Response List Subject Documents Professor Subjects  Subject  Documents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "professor"
        ],
        "summary": "Upload Document",
        "description": "Upload a document for a specific subject.\n\nThe document will be stored in the RAG service and optionally indexed\nfor semantic search.",
        "operationId": "upload_document_professor_subjects__subject__documents_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_professor_subjects__subject__documents_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/professor/subjects/{subject}/documents/{file_path}": {
      "delete": {
        "tags": [
          "professor"
        ],
        "summary": "Delete Document",
        "description": "Delete a document from a specific subject.\n\nThis removes the file from the RAG service storage.",
        "operationId": "delete_document_professor_subjects__subject__documents__file_path__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject"
            }
          },
          {
            "name": "file_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/professor/stats": {
      "get": {
        "tags": [
          "professor"
        ],
        "summary": "Get Dashboard Stats",
        "description": "Get dashboard statistics for the professor.\n\nIncludes:\n- Total students across all subjects\n- Total sessions in their subjects\n- Session activity over the last 7 days\n- Per-subject statistics",
        "operationId": "get_dashboard_stats_professor_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardStats"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_document_professor_subjects__subject__documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "tipo_documento": {
            "type": "string",
            "title": "Tipo Documento",
            "default": "teoria"
          },
          "auto_index": {
            "type": "boolean",
            "title": "Auto Index",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_professor_subjects__subject__documents_post"
      },
...
      "DashboardStats": {
        "properties": {
          "total_students": {
            "type": "integer",
            "title": "Total Students"
          },
          "total_sessions": {
            "type": "integer",
            "title": "Total Sessions"
          },
          "total_documents": {
            "type": "integer",
            "title": "Total Documents"
          },
          "subjects": {
            "items": {
              "$ref": "#/components/schemas/SubjectStats"
            },
            "type": "array",
            "title": "Subjects"
          },
          "sessions_last_7_days": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Sessions Last 7 Days"
          }
        },
        "type": "object",
        "required": [
          "total_students",
          "total_sessions",
          "total_documents",
          "subjects",
          "sessions_last_7_days"
        ],
        "title": "DashboardStats"
      },
      "DocumentInfo": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "size_kb": {
            "type": "number",
            "title": "Size Kb"
          },
          "tipo_documento": {
            "type": "string",
            "title": "Tipo Documento"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "path",
          "size_kb",
          "tipo_documento"
        ],
        "title": "DocumentInfo"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StudentInfo": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "username",
          "email"
        ],
        "title": "StudentInfo"
      },
      "SubjectInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "student_count": {
            "type": "integer",
            "title": "Student Count"
          },
          "document_count": {
            "type": "integer",
            "title": "Document Count"
          }
        },
        "type": "object",
        "required": [
          "name",
          "student_count",
          "document_count"
        ],
        "title": "SubjectInfo"
      },
      "SubjectStats": {
        "properties": {
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "session_count": {
            "type": "integer",
            "title": "Session Count"
          },
          "message_count": {
            "type": "integer",
            "title": "Message Count"
          }
        },
        "type": "object",
        "required": [
          "subject",
          "session_count",
          "message_count"
        ],
        "title": "SubjectStats"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  }
}